# datingapp-2025-main/API/openapi/openapi.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Halo API",
    "description": "API cœur de Halo (membres, likes/matchs, messagerie, prompts).",
    "version": "v1"
  },
  "paths": {
...
    "/api/Messages/thread/{recipientId}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "recipientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Messages/{id}": {
      "delete": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MessageDto": {
        "required": [
          "content",
          "id",
          "recipientDisplayName",
          "recipientId",
          "senderDisplayName",
          "senderId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "senderId": {
            "type": "string",
            "nullable": true
          },
          "senderDisplayName": {
            "type": "string",
            "nullable": true
          },
          "senderImageUrl": {
            "type": "string",
            "nullable": true
          },
          "recipientId": {
            "type": "string",
            "nullable": true
          },
          "recipientDisplayName": {
            "type": "string",
            "nullable": true
          },
          "recipientImageUrl": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "dateRead": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSent": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Entrez le token JWT obtenu via /api/account/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}